requirements.txt
ultralytics
pillow
geopy
flask
beautifulsoup4
requests
pytesseract
